implemented a better display fn, modularized
